requirements.txt
contextlib2==0.5.5
Django==1.10.5
django-colorfield==0.1.10
django-extensions==1.7.3
django-filter==1.1.0
django-widget-tweaks==1.4.2
feedparser==5.2.1
psycopg2==2.7.3.2
pytz==2016.7
raven==6.1.0
requests==2.6.0
six==1.10.0
